Fix tester feedback: onboarding, paywall, meal planner, profile cleanup
- Fix cookbook onboarding page 3.9px right overflow (Flexible wrapper)
- Fix progress indicator order: Features → Children → You
- Add account transition page between parent setup and signup
- Fix paywall not showing after signup (goNamedAuth → goNamed on all 3 auth paths)
- Add confirmation snackbar when meal planner days are updated
- Remove temp Upload Activities button from profile
- Add cookbook and todo list pages to onboarding feature walkthrough
- Remove old add-child redirect from home page
- Various meal planner and recipe UI improvements
- iOS Share Extension and deep link handler updates

+++ lib/v2/week_plan/edite_add_meal/edite_add_meal_model.dart
@@ -90,11 +90,10 @@ class EditeAddMealModel extends FlutterFlowModel<EditeAddMealWidget> {
   FocusNode? textFieldFocusNode3;
   TextEditingController? textController3;
   String? Function(BuildContext, String?)? textController3Validator;
-  // Cook time is optional - no validation required
+  // Cook time is optional - accepts integers or decimals
   String? _textController3Validator(BuildContext context, String? val) {
-    // Optional field - if provided, validate it's a number
     if (val != null && val.isNotEmpty) {
-      if (int.tryParse(val) == null) {
+      if (double.tryParse(val) == null) {
         return 'Please enter a valid number';
       }
     }
@@ -105,11 +104,10 @@ class EditeAddMealModel extends FlutterFlowModel<EditeAddMealWidget> {
   FocusNode? textFieldFocusNode4;
   TextEditingController? textController4;
   String? Function(BuildContext, String?)? textController4Validator;
-  // Prep time is optional - no validation required
+  // Prep time is optional - accepts integers or decimals
   String? _textController4Validator(BuildContext context, String? val) {
-    // Optional field - if provided, validate it's a number
     if (val != null && val.isNotEmpty) {
-      if (int.tryParse(val) == null) {
+      if (double.tryParse(val) == null) {
         return 'Please enter a valid number';
       }
     }
